Implemented an optimization in SSIM.ssim_value.
When comparing many images against many other images, it can save a lot of
time to cache SSIM objects.  With that in mind I've modified the
SSIM.ssim_value() method to recognize when it is passed an SSIM object rather
than a regular PIL Image object, and in that case skip the process of building
a new SSIM object for it.  Now to compare an image to another one that already
has an SSIM object created for it, we just pass the img1 member of the second
SSIM object to the ssim_value method of the first SSIM object.