Normalize gaussian kernel
Resolves #9.

The discrete gaussian kernel created in this helper function does
not perfectly sum up to 1. Especially when sigma < 1, the sum of
the kernel will be larger than 1, at around 1.01.

Trying to compute the variance with the algebraic formula creates
a second term that is larger than the first term with a kernel
like this. The variance becomes negative, and the computed SSIM
overflows.